.gitlab-ci.yml: Add FreeBSD CI configuration
This enables running tests on FreeBSD on custom GitLab CI runners that
accept the tag 'freebsd'. FreeBSD Jails is required to be configured on
the host machine.

Closes https://gitlab.com/coala/package_manager/issues/174

+++ .ci/deps.sh
@@ -5,12 +5,17 @@
 $PYTHON -m pip install --upgrade pip setuptools
 npm install -g bower
 curl -fsSL -o haskellstack.sh https://get.haskellstack.org/
-curl -fsSL -o miniconda.sh \
-     http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
-chmod +x haskellstack.sh miniconda.sh
-./haskellstack.sh
-stack setup --resolver ghc-8.4.2
-./miniconda.sh -b -p "$HOME/miniconda"
+if [ -f dependency_management/requirements/HaskellRequirement.py ]; then
+  chmod +x haskellstack.sh
+  ./haskellstack.sh
+  stack setup --resolver ghc-8.4.2
+fi
+if [ -f dependency_management/requirements/CondaRequirement.py ]; then
+  curl -fsSL -o miniconda.sh \
+       http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
+  chmod +x miniconda.sh
+  ./miniconda.sh -b -p "$HOME/miniconda"
+fi
 go get -u github.com/gpmgo/gopm
 go install github.com/gpmgo/gopm
 go get -u gopkg.in/alecthomas/gometalinter.v2
